@@ -1,17 +1,9 @@
 """The dataloader is a primitive for efficiently loading data from a dataset.
 
 For the most part, just using the DataLoader fluent builder is all you should need.
-However, if full customization is needed, the specific dataloader can be used directly
-by importing one of the following:
-
-- NaiveDataLoader: The naive dataloader is a simple dataloader that loads data in a
-  single thread. It is not recommended to use this dataloader directly, but it can
-  be useful for debugging or for simple use cases.
-- DistributedDataLoader: The distributed dataloader is a dataloader that leverages
-  concurrent futures to load data in parallel for a single node and can be combined
-  with a Mesh and PartitionSpec to load data across multiple nodes, intelligently
-  sharding the data across the nodes.
+However, if full customization is needed, the specific dataloader can be used directly.
 """
 
-from loadax.dataloader.builder import DataLoader as DataLoader
-from loadax.dataloader.naive import NaiveDataLoader as NaiveDataLoader
+from loadax.dataloader.builder import DataloaderBuilder as DataloaderBuilder
+from loadax.dataloader.loader import Dataloader as Dataloader
+from loadax.dataloader.naive import NaiveDataloader as NaiveDataloader
