@@ -1,10 +1,10 @@
-from loadax import Batcher, DataLoader, InMemoryDataset
+from loadax import Batcher, DataloaderBuilder, InMemoryDataset
 
 
 def test_single_batch_loader():
     dataset = InMemoryDataset([1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
     batcher = Batcher(lambda x: x)
-    loader = DataLoader(batcher).workers(0).batch_size(2).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(2).build(dataset)
 
     batches = list(loader)
 
@@ -15,7 +15,7 @@ def test_single_batch_loader():
 def test_batching():
     dataset = InMemoryDataset([1, 2, 3, 4, 5, 6])
     batcher = Batcher(lambda items: sum(items))
-    dataloader = DataLoader(batcher).workers(0).batch_size(2).build(dataset)
+    dataloader = DataloaderBuilder(batcher).workers(0).batch_size(2).build(dataset)
 
     values = list(dataloader)
 
@@ -25,7 +25,7 @@ def test_batching():
 def test_dataloader_iteration():
     dataset = InMemoryDataset([1, 2, 3, 4, 5, 6])
     batcher = Batcher(lambda items: sum(items))
-    dataloader = DataLoader(batcher).workers(0).batch_size(2).build(dataset)
+    dataloader = DataloaderBuilder(batcher).workers(0).batch_size(2).build(dataset)
 
     values = list(dataloader)
     assert values == [3, 7, 11]
@@ -41,7 +41,7 @@ def test_dataloader_iteration():
 def test_empty_dataset():
     dataset = InMemoryDataset([])
     batcher = Batcher(lambda x: x)
-    loader = DataLoader(batcher).workers(0).batch_size(2).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(2).build(dataset)
 
     batches = list(loader)
 
@@ -51,7 +51,7 @@ def test_empty_dataset():
 def test_single_item_dataset():
     dataset = InMemoryDataset([1])
     batcher = Batcher(lambda x: x)
-    loader = DataLoader(batcher).workers(0).batch_size(2).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(2).build(dataset)
 
     batches = list(loader)
 
@@ -61,7 +61,7 @@ def test_single_item_dataset():
 def test_non_divisible_batch_size():
     dataset = InMemoryDataset([1, 2, 3, 4, 5])
     batcher = Batcher(lambda x: x)
-    loader = DataLoader(batcher).workers(0).batch_size(2).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(2).build(dataset)
 
     batches = list(loader)
 
@@ -72,7 +72,7 @@ def test_non_divisible_batch_size():
 def test_large_batch_size():
     dataset = InMemoryDataset([1, 2, 3])
     batcher = Batcher(lambda x: x)
-    loader = DataLoader(batcher).workers(0).batch_size(10).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(10).build(dataset)
 
     batches = list(loader)
 
@@ -83,7 +83,7 @@ def test_large_batch_size():
 def test_small_batch_size():
     dataset = InMemoryDataset([1, 2, 3, 4, 5])
     batcher = Batcher(lambda x: x)
-    loader = DataLoader(batcher).workers(0).batch_size(1).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(1).build(dataset)
 
     batches = list(loader)
 
@@ -94,7 +94,7 @@ def test_small_batch_size():
 def test_dataloader_different_data_types():
     dataset = InMemoryDataset([1, "a", 3.0, {"key": "value"}, True])
     batcher = Batcher(lambda x: x)
-    loader = DataLoader(batcher).workers(0).batch_size(2).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(2).build(dataset)
 
     batches = list(loader)
 
@@ -105,7 +105,7 @@ def test_dataloader_different_data_types():
 def test_dataloader_custom_batch_function():
     dataset = InMemoryDataset([1, 2, 3, 4, 5, 6])
     batcher = Batcher(lambda items: [item * 2 for item in items])
-    loader = DataLoader(batcher).workers(0).batch_size(3).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(3).build(dataset)
 
     batches = list(loader)
 
@@ -126,7 +126,7 @@ class CustomDataset(InMemoryDataset):
 def test_dataloader_custom_dataset():
     dataset = CustomDataset([1, 2, 3, 4])
     batcher = Batcher(lambda x: x)
-    loader = DataLoader(batcher).workers(0).batch_size(2).build(dataset)
+    loader = DataloaderBuilder(batcher).workers(0).batch_size(2).build(dataset)
 
     batches = list(loader)
 
