@@ -1,12 +1,12 @@
 import time
 
-from loadax import Batcher, DataLoader, InMemoryDataset
+from loadax import Batcher, DataloaderBuilder, InMemoryDataset
 
 
 def test_multiprocessing_dataloader():
     dataset = InMemoryDataset([1, 2, 3, 4, 5])
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(2).workers(2).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(2).workers(2).prefetch(2)
     dataloader = builder.build(dataset)
 
     result = []
@@ -19,7 +19,7 @@ def test_multiprocessing_dataloader():
 def test_multiprocessing_dataloader_batch_size():
     dataset = InMemoryDataset(list(range(10)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(3).workers(2).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(3).workers(2).prefetch(2)
     dataloader = builder.build(dataset)
 
     batches = list(dataloader)
@@ -34,7 +34,7 @@ def test_multiprocessing_dataloader_batch_size():
 def test_multiprocessing_dataloader_large_dataset():
     dataset = InMemoryDataset(list(range(100)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(10).workers(4).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(10).workers(4).prefetch(2)
     dataloader = builder.build(dataset)
 
     result = []
@@ -47,7 +47,7 @@ def test_multiprocessing_dataloader_large_dataset():
 def test_multiprocessing_dataloader_empty_dataset():
     dataset = InMemoryDataset([])
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(2).workers(2).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(2).workers(2).prefetch(2)
     dataloader = builder.build(dataset)
 
     result = list(dataloader)
@@ -58,7 +58,7 @@ def test_multiprocessing_dataloader_empty_dataset():
 def test_multiprocessing_dataloader_single_element():
     dataset = InMemoryDataset([1])
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(2).workers(2).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(2).workers(2).prefetch(2)
     dataloader = builder.build(dataset)
 
     result = list(dataloader)
@@ -69,7 +69,7 @@ def test_multiprocessing_dataloader_single_element():
 def test_multiprocessing_dataloader_non_divisible_batch_size():
     dataset = InMemoryDataset(list(range(7)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(3).workers(2).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(3).workers(2).prefetch(2)
     dataloader = builder.build(dataset)
 
     batches = list(dataloader)
@@ -83,7 +83,7 @@ def test_multiprocessing_dataloader_non_divisible_batch_size():
 def test_multiprocessing_dataloader_multiple_workers_large_dataset():
     dataset = InMemoryDataset(list(range(1000)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(100).workers(10).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(100).workers(10).prefetch(2)
     dataloader = builder.build(dataset)
 
     result = []
@@ -96,7 +96,7 @@ def test_multiprocessing_dataloader_multiple_workers_large_dataset():
 def test_multiprocessing_dataloader_worker_handling():
     dataset = InMemoryDataset(list(range(20)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(4).workers(3).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(4).workers(3).prefetch(2)
     dataloader = builder.build(dataset)
 
     batches = list(dataloader)
@@ -112,7 +112,7 @@ def test_multiprocessing_dataloader_worker_handling():
 def test_multiprocessing_dataloader_different_prefetch_factors():
     dataset = InMemoryDataset(list(range(20)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(5).workers(2).prefetch(4)
+    builder = DataloaderBuilder(batcher).batch_size(5).workers(2).prefetch(4)
     dataloader = builder.build(dataset)
 
     batches = list(dataloader)
@@ -127,7 +127,7 @@ def test_multiprocessing_dataloader_different_prefetch_factors():
 def test_multiprocessing_dataloader_varying_data_types():
     dataset = InMemoryDataset([1, "a", 3.0, True, None])
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(2).workers(2).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(2).workers(2).prefetch(2)
     dataloader = builder.build(dataset)
 
     batches = list(dataloader)
@@ -140,7 +140,7 @@ def test_multiprocessing_dataloader_varying_data_types():
 def test_multiprocessing_dataloader_large_number_of_small_batches():
     dataset = InMemoryDataset(list(range(50)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(1).workers(5).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(1).workers(5).prefetch(2)
     dataloader = builder.build(dataset)
 
     result = []
@@ -153,7 +153,7 @@ def test_multiprocessing_dataloader_large_number_of_small_batches():
 def test_multiprocessing_dataloader_odd_elements_and_batch_size():
     dataset = InMemoryDataset(list(range(15)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(4).workers(2).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(4).workers(2).prefetch(2)
     dataloader = builder.build(dataset)
 
     batches = list(dataloader)
@@ -177,7 +177,7 @@ class SlowDataset(InMemoryDataset):
 def test_multiprocessing_dataloader_slow_dataset():
     dataset = SlowDataset(list(range(10)))
     batcher = Batcher(lambda x: x)
-    builder = DataLoader(batcher).batch_size(2).workers(2).prefetch(2)
+    builder = DataloaderBuilder(batcher).batch_size(2).workers(2).prefetch(2)
     dataloader = builder.build(dataset)
 
     result = []
